+++ src/presentation/pages/login/login.spec.tsx
@@ -56,27 +56,11 @@ const simulateValidSubmit = async (
   email = faker.internet.email(),
   password = faker.internet.password()
 ): Promise<void> => {
-  populateEmailField(sut, email);
-  populatePasswordField(sut, password);
+  Helper.populateField(sut, 'email', email);
+  Helper.populateField(sut, 'password', password);
   const form = sut.getByTestId('form');
   fireEvent.submit(form);
   await waitFor(() => form);
-};
-
-const populateEmailField = (sut: RenderResult, email = faker.internet.email()): void => {
-  const emailInput = sut.getByTestId('email');
-  fireEvent.input(emailInput, {
-    target: { value: email },
-  });
-};
-const populatePasswordField = (
-  sut: RenderResult,
-  password = faker.internet.password()
-): void => {
-  const passwordInput = sut.getByTestId('password');
-  fireEvent.input(passwordInput, {
-    target: { value: password },
-  });
 };
 
 const testElementExists = (sut: RenderResult, fieldName: string): void => {
@@ -102,33 +86,33 @@ describe('Login Component', () => {
   test('Should show emailError if validation fails', () => {
     const validationError = faker.word.words();
     const { sut } = makeSut({ validationError });
-    populateEmailField(sut);
+    Helper.populateField(sut, 'email');
     Helper.testStatusForField(sut, 'email', validationError);
   });
 
   test('Should show passwordError if validation fails', () => {
     const validationError = faker.word.words();
     const { sut } = makeSut({ validationError });
-    populatePasswordField(sut);
+    Helper.populateField(sut, 'password');
     Helper.testStatusForField(sut, 'password', validationError);
   });
 
   test('Should show valid email state if validation succeeds', () => {
     const { sut } = makeSut();
-    populateEmailField(sut);
+    Helper.populateField(sut, 'email');
     Helper.testStatusForField(sut, 'email');
   });
 
   test('Should show valid password state if validation succeeds', () => {
     const { sut } = makeSut();
-    populatePasswordField(sut);
+    Helper.populateField(sut, 'password');
     Helper.testStatusForField(sut, 'password');
   });
 
   test('Should enable submit button is form is valid', () => {
     const { sut } = makeSut();
-    populateEmailField(sut);
-    populatePasswordField(sut);
+    Helper.populateField(sut, 'email');
+    Helper.populateField(sut, 'password');
     Helper.testButtonIsDisabled(sut, 'submit', false);
   });
