+++ src/presentation/test/form-helper.tsx
@@ -1,4 +1,5 @@
-import { RenderResult } from '@testing-library/react';
+import { faker } from '@faker-js/faker';
+import { RenderResult, fireEvent } from '@testing-library/react';
 
 export const testChildCount = (
   sut: RenderResult,
@@ -24,4 +25,15 @@ export const testStatusForField = (
   const fieldStatus = sut.getByTestId(`${fieldName}-status`);
   expect(fieldStatus.title).toBe(validationError || 'OK');
   expect(fieldStatus.textContent).toBe(validationError ? '🔴' : '🟢');
+};
+
+export const populateField = (
+  sut: RenderResult,
+  fieldName: string,
+  value = faker.word.verb()
+): void => {
+  const input = sut.getByTestId(fieldName);
+  fireEvent.input(input, {
+    target: { value },
+  });
 };
